chore(virtctl): Move stop command to own file
