+++ pkg/virtctl/vm/vm.go
@@ -38,7 +38,6 @@ import (
 )
 
 const (
-	COMMAND_STOP           = "stop"
 	COMMAND_RESTART        = "restart"
 	COMMAND_MIGRATE_CANCEL = "migrate-cancel"
 	COMMAND_GUESTOSINFO    = "guestosinfo"
@@ -72,25 +71,6 @@ var (
 	dryRun       bool
 	cache        string
 )
-
-func NewStopCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "stop (VM)",
-		Short:   "Stop a virtual machine.",
-		Example: usage(COMMAND_STOP),
-		Args:    templates.ExactArgs("stop", 1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			c := Command{command: COMMAND_STOP, clientConfig: clientConfig}
-			return c.Run(args)
-		},
-	}
-
-	cmd.Flags().BoolVar(&forceRestart, forceArg, false, "--force=false: Only used when grace-period=0. If true, immediately remove VMI pod from API and bypass graceful deletion. Note that immediate deletion of some resources may result in inconsistency or data loss and requires confirmation.")
-	cmd.Flags().Int64Var(&gracePeriod, gracePeriodArg, notDefinedGracePeriod, "--grace-period=-1: Period of time in seconds given to the VMI to terminate gracefully. Can only be set to 0 when --force is true (force deletion). Currently only setting 0 is supported.")
-	cmd.Flags().BoolVar(&dryRun, dryRunArg, false, dryRunCommandUsage)
-	cmd.SetUsageTemplate(templates.UsageTemplate())
-	return cmd
-}
 
 func NewRestartCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
 	cmd := &cobra.Command{
@@ -374,25 +354,6 @@ func (o *Command) Run(args []string) error {
 		fmt.Printf("Dry Run execution\n")
 	}
 	switch o.command {
-	case COMMAND_STOP:
-		if gracePeriodIsSet(gracePeriod) && forceRestart == false {
-			return fmt.Errorf("Can not set gracePeriod without --force=true")
-		}
-		if forceRestart {
-			if gracePeriodIsSet(gracePeriod) {
-				err = virtClient.VirtualMachine(namespace).ForceStop(context.Background(), vmiName, &v1.StopOptions{GracePeriod: &gracePeriod, DryRun: dryRunOption})
-				if err != nil {
-					return fmt.Errorf("Error force stopping VirtualMachine, %v", err)
-				}
-			} else if !gracePeriodIsSet(gracePeriod) {
-				return fmt.Errorf("Can not force stop without gracePeriod")
-			}
-			break
-		}
-		err = virtClient.VirtualMachine(namespace).Stop(context.Background(), vmiName, &v1.StopOptions{DryRun: dryRunOption})
-		if err != nil {
-			return fmt.Errorf("Error stopping VirtualMachine %v", err)
-		}
 	case COMMAND_RESTART:
 		if gracePeriod != notDefinedGracePeriod && forceRestart == false {
 			return fmt.Errorf("Can not set gracePeriod without --force=true")
